+++ VoxEngine/Game/World/Chunk/Light.h
@@ -22,12 +22,12 @@ struct LightNode
 {
 	uint8_t x : 4, y : 4, z : 4;
 
-	LightNode(int x, int y, int z);
+	LightNode(uint8_t x, uint8_t y, uint8_t z);
 };
 
 struct LightRemovalNode
 {
 	uint8_t x : 4, y : 4, z : 4, lightLevel : 4;
 
-	LightRemovalNode(int x, int y, int z, uint8_t lightLevel);
+	LightRemovalNode(uint8_t x, uint8_t y, uint8_t z, uint8_t lightLevel);
 };
